+++ packages/graphql/src/schema/participant.ts
@@ -1,17 +1,20 @@
 import * as DB from '@klicker-uzh/prisma'
 import builder from '../builder'
-import { ParticipantAchievementInstance } from './achievement'
-import {
-  CourseRef,
-  GroupLeaderboardEntry,
+import { Achievement, ParticipantAchievementInstance } from './achievement'
+import type {
   ICourse,
   IGroupLeaderboardEntry,
   ILeaderboardEntry,
   ILeaderboardStatistics,
+} from './course'
+import {
+  CourseRef,
+  GroupLeaderboardEntry,
   LeaderboardEntry,
   LeaderboardStatistics,
 } from './course'
-import { IQuestionInstance, QuestionInstanceRef } from './question'
+import type { IQuestionInstance } from './question'
+import { QuestionInstanceRef } from './question'
 
 export const AvatarSettingsInput = builder.inputType('AvatarSettingsInput', {
   fields: (t) => ({
@@ -51,12 +54,31 @@ export const SubscriptionObjectInput = builder.inputType(
   }
 )
 
+export interface ILevel extends DB.Level {
+  nextLevel?: ILevel
+}
+export const LevelRef = builder.objectRef<ILevel>('Level')
+export const Level = LevelRef.implement({
+  fields: (t) => ({
+    id: t.exposeInt('id'),
+    index: t.exposeInt('index'),
+    requiredXp: t.exposeInt('requiredXp'),
+    name: t.exposeString('name', { nullable: true }),
+    avatar: t.exposeString('avatar', { nullable: true }),
+    nextLevel: t.expose('nextLevel', {
+      type: LevelRef,
+      nullable: true,
+    }),
+  }),
+})
+
 export interface IParticipant extends DB.Participant {
   rank?: number
   score?: number
   isSelf?: boolean
   achievements?: DB.ParticipantAchievementInstance[]
   participantGroups?: IParticipantGroup[]
+  levelData?: ILevel
 }
 export const ParticipantRef = builder.objectRef<IParticipant>('Participant')
 export const Participant = ParticipantRef.implement({
@@ -72,6 +94,7 @@ export const Participant = ParticipantRef.implement({
 
     xp: t.exposeInt('xp'),
     level: t.exposeInt('level'),
+    levelData: t.expose('levelData', { type: LevelRef, nullable: true }),
 
     participantGroups: t.expose('participantGroups', {
       type: [ParticipantGroupRef],
@@ -230,3 +253,21 @@ export const LeaveCourseParticipation = LeaveCourseParticipationRef.implement({
     }),
   }),
 })
+
+export interface IParticipantWithAchievements {
+  participant: IParticipant
+  achievements: DB.Achievement[]
+}
+export const ParticipantWithAchievementsRef =
+  builder.objectRef<IParticipantWithAchievements>('ParticipantWithAchievements')
+export const ParticipantWithAchievements =
+  ParticipantWithAchievementsRef.implement({
+    fields: (t) => ({
+      participant: t.expose('participant', {
+        type: ParticipantRef,
+      }),
+      achievements: t.expose('achievements', {
+        type: [Achievement],
+      }),
+    }),
+  })