+++ packages/graphql/src/services/learningElements.ts
@@ -1,4 +1,5 @@
 import {
+  computeAwardedXp,
   gradeQuestionKPRIM,
   gradeQuestionMC,
   gradeQuestionNumerical,
@@ -6,6 +7,7 @@ import {
 } from '@klicker-uzh/grading'
 import {
   Attachment,
+  LearningElementStatus,
   OrderType,
   Question,
   QuestionInstance,
@@ -22,6 +24,7 @@ import { shuffle } from '../util'
 import { prepareInitialInstanceResults, processQuestionData } from './sessions'
 
 const POINTS_AWARD_TIMEFRAME_DAYS = 6
+const XP_AWARD_TIMEFRAME_DAYS = 1
 
 type QuestionResponse = {
   choices?: number[] | null
@@ -68,6 +71,10 @@ function evaluateQuestionResponse(
             pointsPercentage !== null
               ? round(pointsPercentage * 10 * (multiplier ?? 1))
               : -1,
+          xp: computeAwardedXp({
+            pointsPercentage,
+            multiplier: multiplier ?? 1,
+          }),
         }
       } else if (data.type === QuestionType.MC) {
         const pointsPercentage = gradeQuestionMC({
@@ -82,6 +89,10 @@ function evaluateQuestionResponse(
             pointsPercentage !== null
               ? round(pointsPercentage * 10 * (multiplier ?? 1))
               : -1,
+          xp: computeAwardedXp({
+            pointsPercentage,
+            multiplier: multiplier ?? 1,
+          }),
         }
       } else {
         const pointsPercentage = gradeQuestionKPRIM({
@@ -96,6 +107,10 @@ function evaluateQuestionResponse(
             pointsPercentage !== null
               ? round(pointsPercentage * 10 * (multiplier ?? 1))
               : -1,
+          xp: computeAwardedXp({
+            pointsPercentage,
+            multiplier: multiplier ?? 1,
+          }),
         }
       }
     }
@@ -114,6 +129,10 @@ function evaluateQuestionResponse(
         feedbacks: [],
         answers: results ?? [],
         score: correct ? correct * 10 * (multiplier ?? 1) : 0,
+        xp: computeAwardedXp({
+          pointsPercentage: correct,
+          multiplier: multiplier ?? 1,
+        }),
       }
     }
 
@@ -141,22 +160,23 @@ export async function respondToQuestionInstance(
       where: { id },
       // if the participant is logged in, fetch the last response of the participant
       // the response will not be counted and will only yield points if not within the past week
-      include: ctx.user?.sub
-        ? {
-            responses: {
-              where: {
-                participant: {
-                  id: ctx.user.sub,
+      include:
+        ctx.user?.sub && ctx.user.role === UserRole.PARTICIPANT
+          ? {
+              responses: {
+                where: {
+                  participant: {
+                    id: ctx.user.sub,
+                  },
                 },
+                take: 1,
               },
-              take: 1,
+            }
+          : {
+              responses: {
+                take: 1,
+              },
             },
-          }
-        : {
-            responses: {
-              take: 1,
-            },
-          },
     })
 
     if (!instance) {
@@ -248,9 +268,13 @@ export async function respondToQuestionInstance(
     updatedInstance.pointsMultiplier
   )
   const score = evaluation?.score || 0
+  const xp = evaluation?.xp || 0
   let pointsAwarded
   let newPointsFrom
   let lastAwardedAt
+  let lastXpAwardedAt
+  let xpAwarded
+  let newXpFrom
   const promises = []
 
   // if the user is logged in and the last response was not within the past 6 days
@@ -274,18 +298,40 @@ export async function respondToQuestionInstance(
         pointsAwarded = 0
       }
 
+      const previousXpResponseOutsideTimeframe =
+        !instance.responses[0].lastXpAwardedAt ||
+        dayjs(instance.responses[0].lastXpAwardedAt).isBefore(
+          dayjs().subtract(XP_AWARD_TIMEFRAME_DAYS, 'days')
+        )
+
+      if (previousXpResponseOutsideTimeframe) {
+        xpAwarded = xp
+      } else {
+        xpAwarded = 0
+      }
+
       lastAwardedAt = previousResponseOutsideTimeframe
         ? new Date()
         : instance.responses[0].lastAwardedAt
+      lastXpAwardedAt = previousXpResponseOutsideTimeframe
+        ? new Date()
+        : instance.responses[0].lastXpAwardedAt
       newPointsFrom = dayjs(lastAwardedAt)
         .add(instance?.resetTimeDays ?? POINTS_AWARD_TIMEFRAME_DAYS, 'days')
         .toDate()
+      newXpFrom = dayjs(lastXpAwardedAt)
+        .add(XP_AWARD_TIMEFRAME_DAYS, 'days')
+        .toDate()
     } else {
       pointsAwarded = score
+      xpAwarded = xp
 
       lastAwardedAt = new Date()
       newPointsFrom = dayjs(lastAwardedAt)
         .add(instance?.resetTimeDays ?? POINTS_AWARD_TIMEFRAME_DAYS, 'days')
+        .toDate()
+      newXpFrom = dayjs(lastAwardedAt)
+        .add(XP_AWARD_TIMEFRAME_DAYS, 'days')
         .toDate()
     }
 
@@ -300,8 +346,10 @@ export async function respondToQuestionInstance(
         create: {
           totalScore: score,
           totalPointsAwarded: pointsAwarded,
+          totalXpAwarded: xpAwarded,
           trialsCount: 1,
           lastAwardedAt,
+          lastXpAwardedAt,
           response,
           participant: {
             connect: { id: ctx.user.sub },
@@ -321,6 +369,7 @@ export async function respondToQuestionInstance(
         update: {
           response,
           lastAwardedAt,
+          lastXpAwardedAt,
           trialsCount: {
             increment: 1,
           },
@@ -330,12 +379,16 @@ export async function respondToQuestionInstance(
           totalPointsAwarded: {
             increment: pointsAwarded,
           },
+          totalXpAwarded: {
+            increment: xpAwarded,
+          },
         },
       }),
       ctx.prisma.questionResponseDetail.create({
         data: {
           score,
           pointsAwarded,
+          xpAwarded,
           response,
           participant: {
             connect: { id: ctx.user.sub },
@@ -354,6 +407,21 @@ export async function respondToQuestionInstance(
         },
       })
     )
+
+    if (typeof xpAwarded === 'number' && xpAwarded > 0) {
+      promises.push(
+        ctx.prisma.participant.update({
+          where: {
+            id: ctx.user.sub,
+          },
+          data: {
+            xp: {
+              increment: xpAwarded,
+            },
+          },
+        })
+      )
+    }
 
     if (typeof pointsAwarded === 'number') {
       promises.push(
@@ -406,6 +474,8 @@ export async function respondToQuestionInstance(
           ...evaluation,
           pointsAwarded,
           newPointsFrom,
+          xpAwarded,
+          newXpFrom,
         }
       : null,
   }
@@ -416,26 +486,37 @@ export async function getLearningElementData(
   ctx: Context
 ) {
   const element = await ctx.prisma.learningElement.findUnique({
-    where: { id },
+    where: {
+      id,
+      OR: [
+        {
+          status: LearningElementStatus.PUBLISHED,
+        },
+        {
+          ownerId: ctx.user?.sub,
+        },
+      ],
+    },
     include: {
       course: true,
       instances: {
         orderBy: {
           questionId: 'asc',
         },
-        include: ctx.user?.sub
-          ? {
-              responses: {
-                where: {
-                  participantId: ctx.user.sub,
+        include:
+          ctx.user?.sub && ctx.user?.role === UserRole.PARTICIPANT
+            ? {
+                responses: {
+                  where: {
+                    participantId: ctx.user.sub,
+                  },
+                  orderBy: {
+                    lastAwardedAt: 'asc',
+                  },
+                  take: 1,
                 },
-                orderBy: {
-                  lastAwardedAt: 'asc',
-                },
-                take: 1,
-              },
-            }
-          : undefined,
+              }
+            : undefined,
       },
     },
   })
@@ -665,4 +746,65 @@ export async function getBookmarksLearningElement(
   })
 
   return participation?.bookmarkedQuestions
+}
+
+interface PublishLearningElementArgs {
+  id: string
+}
+
+export async function publishLearningElement(
+  { id }: PublishLearningElementArgs,
+  ctx: ContextWithUser
+) {
+  const learningElement = await ctx.prisma.learningElement.update({
+    where: {
+      id,
+    },
+    data: {
+      status: LearningElementStatus.PUBLISHED,
+    },
+    include: {
+      instances: true,
+    },
+  })
+
+  return {
+    ...learningElement,
+    numOfInstances: learningElement.instances.length,
+  }
+}
+
+interface DeleteLearningElementArgs {
+  id: string
+}
+
+export async function deleteLearningElement(
+  { id }: DeleteLearningElementArgs,
+  ctx: ContextWithUser
+) {
+  try {
+    const deletedItem = await ctx.prisma.learningElement.delete({
+      where: {
+        id,
+        status: LearningElementStatus.DRAFT,
+      },
+    })
+
+    ctx.emitter.emit('invalidate', {
+      typename: 'LearningElement',
+      id,
+    })
+
+    return deletedItem
+  } catch (e) {
+    // TODO: resolve type issue by first testing for prisma error
+    if (e?.code === 'P2025') {
+      console.log(
+        'The learning element is not in draft status and cannot be deleted.'
+      )
+      return null
+    }
+
+    throw e
+  }
 }