+++ packages/prisma/src/prisma/migrations/20230301185232_draft_as_default/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "LearningElement" ALTER COLUMN "status" SET DEFAULT 'DRAFT';
+
+-- AlterTable
+ALTER TABLE "MicroSession" ALTER COLUMN "status" SET DEFAULT 'DRAFT';